refactor:Separate method from DrawManager
변경 사항 : DrawManager에서 GameDrawManager, GameSettingDrawManager, ScoreDrawManager, TwoPlayerDrawManager에 메소드 분리하고, DrawManager에서 주석 처리했습니다. 이후 각 Screen 클래스에서 해당 메소드들을 각 DrawManager에서 불러오도록 수정하고, 게임이 돌아가는 것까지 확인했습니다.

CSE2024SDP-187
